+++ bluetooth/nrf52/NRF52RxTx.cpp
@@ -1,6 +1,6 @@
 #include "NRF52RxTx.h"
 
-void NRF52RxTx::send(char payload[], size_t len)
+void NRF52RxTx::send(char payload[], int len)
 {
     //Tim's garb daddy code
 }
